+++ src/Config/index.ts
@@ -1,34 +1,46 @@
 /**
  *  @FileID          Config\index.ts
  *  @Description     configuration settings for the project.
- *  @Author          @MeetBhingradiya
+ *  @Author          Meet Bhingradiya (@MeetBhingradiya)
  *  
  *  -----------------------------------------------------------------------------
- *  Copyright (c) 2025 Meet Bhingradiya
+ *  
+ *  Copyright (c) 2021 - 2025 Meet Bhingradiya.
  *  All rights reserved.
  *  
- *  This file is part of the @MeetBhingradiya's Portfolio project and is protected under copyright
- *  law. Unauthorized copying of this file, via any medium, is strictly prohibited
- *  without explicit permission from the author.
+ *  This file is a proprietary component of Meet Bhingradiya's Portfolio project
+ *  and is protected under applicable copyright and intellectual property laws.
+ *  Unauthorized use, reproduction, distribution, folks, or modification of this file,
+ *  via any medium, is strictly prohibited without prior written consent from the
+ *  author or the organization.
  *  
  *  -----------------------------------------------------------------------------
+ *  
+ *  Notice: GitHub® is a registered trademark of Microsoft Corporation. This project 
+ *  is not affiliated with, endorsed by, or in any way associated with GitHub or 
+ *  Microsoft Corporation.
+ *  
+ *  -----------------------------------------------------------------------------
+ *  Last sUpdated on Version: 1.0.8
+ *  -----------------------------------------------------------------------------
  *  @created 13/01/25 11:34 AM IST (Kolkata +5:30 UTC)
- *  @modified 19/01/25 9:44 AM IST (Kolkata +5:30 UTC)
+ *  @modified 22/01/25 11:34 AM IST (Kolkata +5:30 UTC)
  */
 
+
 const Config = {
-    version: "1.0.7",
+    version: "1.0.8",
     releasedate: "2025-1-30",
     visiblebranch: "Development",
     isHomeReleased: true,
     Environment: process.env.NODE_ENV,
     GoogleADS: false,
     WhiteListedDomains: [
+        "meetbhingradiya.tech",
+        // "meetbhingradiya.vercel.app",
         // "admin.meetbhingradiya.tech",
         "stage.meetbhingradiya.tech",
         "dev.meetbhingradiya.tech",
-        "meetbhingradiya.tech",
-        // "meetbhingradiya.vercel.app",
         // "dev-meetbhingradiya.vercel.app",
         // "stage-meetbhingradiya.vercel.app",
         // "admin-meetbhingradiya.vercel.app"