+++ src/Models/VerificationCodes.ts
@@ -1,25 +1,37 @@
 /**
  *  @FileID          Models\VerificationCodes.ts
  *  @Description     Currently, there is no description available.
- *  @Author          @MeetBhingradiya
+ *  @Author          Meet Bhingradiya (@MeetBhingradiya)
  *  
  *  -----------------------------------------------------------------------------
- *  Copyright (c) 2025 Meet Bhingradiya
+ *  
+ *  Copyright (c) 2021 - 2025 Meet Bhingradiya.
  *  All rights reserved.
  *  
- *  This file is part of the @MeetBhingradiya's Portfolio project and is protected under copyright
- *  law. Unauthorized copying of this file, via any medium, is strictly prohibited
- *  without explicit permission from the author.
+ *  This file is a proprietary component of Meet Bhingradiya's Portfolio project
+ *  and is protected under applicable copyright and intellectual property laws.
+ *  Unauthorized use, reproduction, distribution, folks, or modification of this file,
+ *  via any medium, is strictly prohibited without prior written consent from the
+ *  author or the organization.
  *  
  *  -----------------------------------------------------------------------------
+ *  
+ *  Notice: GitHub® is a registered trademark of Microsoft Corporation. This project 
+ *  is not affiliated with, endorsed by, or in any way associated with GitHub or 
+ *  Microsoft Corporation.
+ *  
+ *  -----------------------------------------------------------------------------
+ *  Last sUpdated on Version: 1.0.8
+ *  -----------------------------------------------------------------------------
  *  @created 13/01/25 11:34 AM IST (Kolkata +5:30 UTC)
- *  @modified 14/01/25 3:22 PM IST (Kolkata +5:30 UTC)
+ *  @modified 22/01/25 11:34 AM IST (Kolkata +5:30 UTC)
  */
 
-import { Schema, Document, model } from 'mongoose';
+
+import mongoose from 'mongoose';
 import { v4 } from 'uuid';
 
-const VerificationCodes_Schema: Schema = new Schema({
+const VerificationCodes_Schema: mongoose.Schema = new mongoose.Schema({
     CodeID: {
         type: String,
         default: v4,
@@ -33,6 +45,15 @@ const VerificationCodes_Schema: Schema = new Schema({
         type: String,
         required: true
     },
+    Type: {
+        type: String,
+        enum: ["Email", "Phone"],
+        required: true
+    },
+    Expiry: {
+        type: Date,
+        default: new Date(Date.now() + 15 * 60000)
+    },
     UserID: {
         type: String,
         required: true
@@ -40,15 +61,16 @@ const VerificationCodes_Schema: Schema = new Schema({
 }, {
     timestamps: true,
     versionKey: true,
-    expires: 600,
     _id: false
 });
 
-export interface IVerificationCodes extends Document {
-    CodeID: string;
-    HashKey: string;
-    Salt: string;
-    UserID: string;
+export interface IVerificationCodes extends mongoose.Document {
+    CodeID: string
+    HashKey: string
+    Salt: string
+    Type: "Email" | "Phone"
+    Expiry: Date
+    UserID: string
 }
 
-export default model<IVerificationCodes>('VerificationCodes', VerificationCodes_Schema);
+export const VerificationCodes_Model: mongoose.Model<IVerificationCodes> = mongoose.models?.VerificationCodes || mongoose.model<IVerificationCodes>("VerificationCodes", VerificationCodes_Schema);