+++ src/Types/Tools.ts
@@ -1,20 +1,32 @@
 /**
  *  @FileID          Types\Tools.ts
  *  @Description     Currently, there is no description available.
- *  @Author          @MeetBhingradiya
+ *  @Author          Meet Bhingradiya (@MeetBhingradiya)
  *  
  *  -----------------------------------------------------------------------------
- *  Copyright (c) 2025 Meet Bhingradiya
+ *  
+ *  Copyright (c) 2021 - 2025 Meet Bhingradiya.
  *  All rights reserved.
  *  
- *  This file is part of the @MeetBhingradiya's Portfolio project and is protected under copyright
- *  law. Unauthorized copying of this file, via any medium, is strictly prohibited
- *  without explicit permission from the author.
+ *  This file is a proprietary component of Meet Bhingradiya's Portfolio project
+ *  and is protected under applicable copyright and intellectual property laws.
+ *  Unauthorized use, reproduction, distribution, folks, or modification of this file,
+ *  via any medium, is strictly prohibited without prior written consent from the
+ *  author or the organization.
  *  
  *  -----------------------------------------------------------------------------
+ *  
+ *  Notice: GitHub® is a registered trademark of Microsoft Corporation. This project 
+ *  is not affiliated with, endorsed by, or in any way associated with GitHub or 
+ *  Microsoft Corporation.
+ *  
+ *  -----------------------------------------------------------------------------
+ *  Last sUpdated on Version: 1.0.8
+ *  -----------------------------------------------------------------------------
  *  @created 13/01/25 11:34 AM IST (Kolkata +5:30 UTC)
- *  @modified 14/01/25 3:22 PM IST (Kolkata +5:30 UTC)
+ *  @modified 22/01/25 11:34 AM IST (Kolkata +5:30 UTC)
  */
+
 
 interface IBookmark {
     // ? Unique identifier for the bookmark
@@ -139,10 +151,8 @@ interface ISuggestion {
 }
 
 interface IState {
-    FilterSuggestions: Array<ISuggestion>
     FilterBookmarks: Array<IBookmark>
     Bookmarks: Array<IBookmark>
-    Suggestions: Array<ISuggestionResponse>
     Query: string
     Settings: {
         isFirstRun: boolean