+++ src/Utils/Axios.ts
@@ -1,20 +1,32 @@
 /**
  *  @FileID          Utils\Axios.ts
  *  @Description     Currently, there is no description available.
- *  @Author          @MeetBhingradiya
+ *  @Author          Meet Bhingradiya (@MeetBhingradiya)
  *  
  *  -----------------------------------------------------------------------------
- *  Copyright (c) 2025 Meet Bhingradiya
+ *  
+ *  Copyright (c) 2021 - 2025 Meet Bhingradiya.
  *  All rights reserved.
  *  
- *  This file is part of the @MeetBhingradiya's Portfolio project and is protected under copyright
- *  law. Unauthorized copying of this file, via any medium, is strictly prohibited
- *  without explicit permission from the author.
+ *  This file is a proprietary component of Meet Bhingradiya's Portfolio project
+ *  and is protected under applicable copyright and intellectual property laws.
+ *  Unauthorized use, reproduction, distribution, folks, or modification of this file,
+ *  via any medium, is strictly prohibited without prior written consent from the
+ *  author or the organization.
  *  
  *  -----------------------------------------------------------------------------
+ *  
+ *  Notice: GitHub® is a registered trademark of Microsoft Corporation. This project 
+ *  is not affiliated with, endorsed by, or in any way associated with GitHub or 
+ *  Microsoft Corporation.
+ *  
+ *  -----------------------------------------------------------------------------
+ *  Last sUpdated on Version: 1.0.8
+ *  -----------------------------------------------------------------------------
  *  @created 13/01/25 11:34 AM IST (Kolkata +5:30 UTC)
- *  @modified 14/01/25 3:22 PM IST (Kolkata +5:30 UTC)
+ *  @modified 22/01/25 11:34 AM IST (Kolkata +5:30 UTC)
  */
+
 
 import axios, { AxiosInstance, AxiosResponse, InternalAxiosRequestConfig } from 'axios';
 import { getCSRFToken } from './getTrace';
@@ -28,7 +40,14 @@ const Axios: AxiosInstance = axios.create({
 
 Axios.interceptors.request.use(
     async (config: InternalAxiosRequestConfig<any>) => {
-        const csrfToken = await getCSRFToken();
+        let csrfToken: any = null;
+        if (localStorage.getItem('trace')) {
+            csrfToken = JSON.parse(localStorage.getItem('trace') || '{}');
+        } else {
+            csrfToken = await getCSRFToken();
+            localStorage.setItem('trace', JSON.stringify(csrfToken));
+        }
+
         if (csrfToken.Status === 1) {   
             config.headers['x-csrf'] = csrfToken.data;
             config.withCredentials = true;
@@ -47,6 +66,11 @@ Axios.interceptors.response.use(
         return response;
     },
     (error: any) => {
+        if (error.response?.status === 403) {
+            localStorage.removeItem('trace');
+            window.location.reload();
+            return Promise.reject(error);
+        }
         return Promise.reject(error);
     }
 );