+++ src/app/Tools/page.tsx
@@ -1,26 +1,38 @@
 /**
  *  @FileID          app\Tools\page.tsx
  *  @Description     Currently, there is no description available.
- *  @Author          @MeetBhingradiya
+ *  @Author          Meet Bhingradiya (@MeetBhingradiya)
  *  
  *  -----------------------------------------------------------------------------
- *  Copyright (c) 2025 Meet Bhingradiya
+ *  
+ *  Copyright (c) 2021 - 2025 Meet Bhingradiya.
  *  All rights reserved.
  *  
- *  This file is part of the @MeetBhingradiya's Portfolio project and is protected under copyright
- *  law. Unauthorized copying of this file, via any medium, is strictly prohibited
- *  without explicit permission from the author.
+ *  This file is a proprietary component of Meet Bhingradiya's Portfolio project
+ *  and is protected under applicable copyright and intellectual property laws.
+ *  Unauthorized use, reproduction, distribution, folks, or modification of this file,
+ *  via any medium, is strictly prohibited without prior written consent from the
+ *  author or the organization.
  *  
  *  -----------------------------------------------------------------------------
+ *  
+ *  Notice: GitHub® is a registered trademark of Microsoft Corporation. This project 
+ *  is not affiliated with, endorsed by, or in any way associated with GitHub or 
+ *  Microsoft Corporation.
+ *  
+ *  -----------------------------------------------------------------------------
+ *  Last sUpdated on Version: 1.0.8
+ *  -----------------------------------------------------------------------------
  *  @created 14/01/25 3:22 PM IST (Kolkata +5:30 UTC)
- *  @modified 14/01/25 3:22 PM IST (Kolkata +5:30 UTC)
+ *  @modified 22/01/25 11:34 AM IST (Kolkata +5:30 UTC)
  */
+
 
 "use client";
 
 import React from "react";
 import { ISearchEngine, ILocale } from "@Types/Tools";
-import type { IBookmark, IState } from "@Types/Tools";
+import type { IBookmark, IState, ISuggestion } from "@Types/Tools";
 import { ToastContainer, toast } from 'react-toastify';
 import { BookmarksDB } from "@Data/Tools";
 import Image from "next/image";
@@ -131,14 +143,21 @@ const StyledMenu = styled((props: MenuProps) => (
 
 const StorageKey = "Tools";
 
+const SearchEnginePresets = {
+    google: `https://www.google.com/search?q=@Query&utm_source=${Config.WhiteListedDomains[0]}`,
+    bing: `https://www.bing.com/search?q=@Query&utm_source=${Config.WhiteListedDomains[0]}`,
+    duckduckgo: `https://duckduckgo.com/?q=@Query&utm_source=${Config.WhiteListedDomains[0]}`,
+    brave: `https://search.brave.com/search?q=@Query&utm_source=${Config.WhiteListedDomains[0]}`,
+    qwant: `https://www.qwant.com/?q=@Query&utm_source=${Config.WhiteListedDomains[0]}`,
+    yahoo: `https://search.yahoo.com/search?p=@Query&utm_source=${Config.WhiteListedDomains[0]}`,
+}
+
 function Tools() {
     // @ States
     const isClient = useWindowCheck();
     const [State, setState] = React.useState<IState>({
-        FilterSuggestions: [],
         FilterBookmarks: [],
         Bookmarks: [...BookmarksDB],
-        Suggestions: [],
         Query: "",
         Settings: {
             isFirstRun: true,
@@ -149,6 +168,7 @@ function Tools() {
             Locale: ILocale.EN
         }
     });
+    const [Suggestions, setSuggestions] = React.useState<Array<ISuggestion>>([]);
     const [windowWidth, setWindowWidth] = React.useState<number>(isClient ? window.innerWidth : 0);
     const [contextMenu, setContextMenu] = React.useState<{
         mouseX: number;
@@ -186,15 +206,36 @@ function Tools() {
             searchInputRef.current.blur();
         }
 
+        // ? On Space Focus on Search
+        if (e.key === " " && searchInputRef.current) {
+            searchInputRef.current.focus();
+        }
+
         // ? On Enter Open First Link
         if (e.key === "Enter") {
             if (State.FilterBookmarks.length > 0 && State.Query.length > 0) {
+                console.log("Opening First Link");
                 window.open(State.FilterBookmarks[0].url, State.Settings.isNewTab ? "_blank" : "_self");
                 setState({
                     ...State,
                     FilterBookmarks: State.Bookmarks,
                     Query: "",
                 });
+                return;
+            }
+
+            if (State.Query.length > 0) {
+                window.open(SearchEngineLinkBuilder(State.Query), State.Settings.isNewTab ? "_blank" : "_self");
+                setState({
+                    ...State,
+                    FilterBookmarks: State.Bookmarks,
+                    Query: "",
+                });
+                return;
+            }
+
+            if (State.Query.length === 0 && searchInputRef.current) {
+                searchInputRef.current.focus();
             }
         }
 
@@ -227,8 +268,14 @@ function Tools() {
         });
     }
 
-    const onQueryChange = (e: any) => {
+    const abortControllerRef = React.useRef<AbortController | null>(null);
+    const onQueryChange = async (e: any) => {
         const query = e.target.value.toLowerCase();
+        const OriginalQuery = e.target.value;
+
+        if (OriginalQuery.trim() === "") {
+            return;
+        }
 
         const isExactMatch = (name: string, keywords: Array<string> = []): boolean => {
             return name.toLowerCase() === query || keywords.some((keyword) => keyword.toLowerCase() === query);
@@ -276,11 +323,71 @@ function Tools() {
 
         setState({
             ...State,
-            Query: query,
+            Query: OriginalQuery,
             FilterBookmarks: filteredBookmarks
         });
+
+        if (query.length > 0) {
+            FetchSuggestions();
+        } else {
+            setSuggestions([]);
+        }
     };
 
+    const FetchSuggestions = async () => {
+        let ProcessedSuggestions: any[] = [];
+
+        if (State.FilterBookmarks.length === 0) {
+            if (abortControllerRef.current) {
+                abortControllerRef.current.abort();
+            }
+
+            const controller = new AbortController();
+            abortControllerRef.current = controller;
+
+            let API = "https://api.suggestions.victr.me/";
+            let _API = new URL(API);
+            _API.searchParams.append("q", State.Query);
+            _API.searchParams.append("l", State.Settings.Locale);
+            _API.searchParams.append("with", State.Settings.SearchEngine);
+
+            try {
+                const response = (await Axios.post(
+                    "/api/cors",
+                    {
+                        body: {
+                            endpoint: _API,
+                            method: "GET",
+                            body: null,
+                            headers: {},
+                        },
+                        method: "POST",
+                    },
+                    { signal: controller.signal }
+                )).data;
+
+                const Suggestions = response.data
+
+                if (!Suggestions) {
+                    return;
+                }
+
+                ProcessedSuggestions = Suggestions.map((suggestion: any) => {
+                    return {
+                        Query: suggestion.text,
+                        Thumbnail: suggestion.image,
+                        Description: suggestion.desc,
+                    };
+                });
+            } catch (error: any) {
+                if (error.name === "CanceledError") {
+                    console.log("Request canceled.");
+                }
+            }
+        }
+
+        setSuggestions(ProcessedSuggestions);
+    };
 
     function handleContextMenu(e: React.MouseEvent<HTMLDivElement>, ID: string) {
         e.preventDefault();
@@ -410,6 +517,11 @@ function Tools() {
         setWindowWidth(window.innerWidth);
     };
 
+    function SearchEngineLinkBuilder(query: string) {
+        let SearchEngine = SearchEnginePresets[State.Settings.SearchEngine as ISearchEngine];
+        return SearchEngine.replace("@Query", query);
+    }
+
     // @Updates
 
     // ? Initial Run & Window Resize, Focus on Search Input Listeners
@@ -481,7 +593,6 @@ function Tools() {
 
     // ? State Sync with Storage
     React.useEffect(() => {
-
         localStorage.setItem(StorageKey, JSON.stringify({
             Booksmarks: State.Bookmarks,
             Settings: {
@@ -542,22 +653,32 @@ function Tools() {
                     State.FilterBookmarks.length === 0 && (
                         <div className="Suggestions">
                             {
-                                State.FilterSuggestions.map((item, index) => (
+                                Suggestions.map((item, index) => (
                                     <div
                                         key={index}
                                         className="Suggestion"
                                         onClick={() => {
-                                            // ! TODO: Implement Link Builder
+                                            window.open(SearchEngineLinkBuilder(item.Query), State.Settings.isNewTab ? "_blank" : "_self");
                                         }}
                                     >
-                                        <h2 className="Thumbnail">{item?.Thumbnail}</h2>
-                                        <p className="Title">{item?.Query}</p>
+                                        {
+                                            item?.Thumbnail && (
+                                                <Image
+                                                    className="Thumbnail"
+                                                    width={64}
+                                                    height={64}
+                                                    src={item?.Thumbnail}
+                                                    alt="Thumbnail"
+                                                />
+                                            )
+                                        }
+                                        <h2 className="Title">{item?.Query}</h2>
                                     </div>
                                 ))
                             }
 
                             {
-                                State.FilterSuggestions.length === 0 && (
+                                Suggestions.length === 0 && (
                                     <div
                                         className="Suggestion"
                                     >